--- main/app/src/classes/Audio.tsx
@@ -1,47 +1,0 @@
-// import React, { Component } from 'react';
-
-// class Audio extends Component {
-//   constructor(props: any) {
-//     super(props);
-//     this.state = {
-//       audio: null
-//     };
-//     this.toggleMicrophone = this.toggleMicrophone.bind(this);
-//   }
-
-//   async getMicrophone() {
-//     const audio = await navigator.mediaDevices.getUserMedia({
-//       audio: true,
-//       video: false
-//     });
-//     this.setState({ audio });
-//   }
-
-//   stopMicrophone() {
-//     audio.getTracks().forEach((track: any) => track.stop());
-//     this.setState({ audio: null });
-//   }
-
-//   toggleMicrophone() {
-//     if (this.state.audio) {
-//       this.stopMicrophone();
-//     } else {
-//       this.getMicrophone();
-//     }
-//   }
-
-//   render() {
-//     return (
-//       <div className="App">
-//         <div className="controls">
-//           <button onClick={this.toggleMicrophone}>
-//             {this.state.audio ? 'Stop microphone' : 'Get microphone input'}
-//           </button>
-//         </div>
-//         {this.state.audio ? <AudioAnalyser audio={this.state.audio} /> : ''}
-//       </div>
-//     );
-//   }
-// }
-
-// export default Audio;
