+++ main/my-app/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "app",
+  "name": "my-app",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -7,12 +7,15 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.57",
-    "@types/react": "^18.2.25",
+    "@types/node": "^16.18.61",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
     "react": "^18.2.0",
-    "react-scripts": "^3.0.1",
-    "react-sound": "^1.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.18.0",
+    "react-scripts": "5.0.1",
     "react-speech-recognition": "^3.10.0",
+    "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -20,6 +23,12 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -34,8 +43,6 @@
     ]
   },
   "devDependencies": {
-    "@types/react-sound": "^1.2.6",
-    "@types/react-speech-recognition": "^3.9.3",
-    "react-router-dom": "^6.17.0"
+    "@types/react-speech-recognition": "^3.9.4"
   }
 }
