+++ main/my-app/public/manifest.json
