+++ main/my-app/src/pages/home.tsx
@@ -6,12 +6,12 @@ import SpeechRecognition, { useSpeechRecognition } from 'react-speech-recognitio
 import "../styles/styles.css";
 import TextToSpeech from '../components/TextToSpeech';
 import AudioCapturePlayback from '../components/AudioPlayback';
-import AudioAnalysis from '../components/AudioAnalysis'
+// import AudioAnalysis from '../components/AudioAnalysis'
 
 const HomePage = () => {
 
-  const navigate = useNavigate();
-  const handleClick = () => navigate('/captions');
+  // const navigate = useNavigate();
+  // const handleClick = () => navigate('/captions');
   
   const [listening, setListening] = useState(false);
 
@@ -57,7 +57,7 @@ const HomePage = () => {
         <TextToSpeech text={transcript} />
 
         <AudioCapturePlayback />
-        <AudioAnalysis />
+        {/* <AudioAnalysis /> */}
       </div>
     </div>
   );
